Move Ed25519 to the encoding package - it's only used by the SolanaProgramDerivedAddress so just put it next to it in the package
